Extract definition query runner partial
